refactor tpu_utils to gcp_utils
